Updated section about ftp validation
Hi. 
-FTP validation-

I added a note about a configuration strategy that works very good for me. In this case i don't have to do any extra configuration when i add extra SAN entries.

I added this mostly because i think there might be something fishy with the acme-plugin in pfsense. Before, the the site I registered did not need to pass the test for if the challenge was reachable from the web (I don't know if this was the right or wrong behavior) but now it must (or at least it needed for me). To keep my administration to a minimum when adding/removing sites and managing access permissions i though this approach was simple. Since the documentation does not explicitly tell if the site needs to be reachable for this validation method i thought it best to provide it non the less.

Sorry about the formatting but it is my first atempt at ReStructuredText.
Hope you find it useful.