Added specifics on how to create these rules
This document pointed me in the right direction but I could not figure out from what was said how to actually do it so it would work. I eventually found this, https://forum.netgate.com/topic/26292/carp-openvpn/3, in the forums which helped decrypt this. I though I better add the specific instructions here to save future searchers hours of work.