Reverted part of 52f0a93, return IPSec PFS suggestion to off
This returns the IPSec phase 2 PFS setting suggestion back to 'off', as it was before my change in 52f0a93.  Did this because unfortunately macOS doesn't actually like that setting, at least not without an MDM config, which is a bit complicated for people doing first time setup with this guide.